+++ tests/test_session.py
@@ -9,6 +9,7 @@ from werobot.utils import to_binary
 
 import pymongo
 import redis
+import os
 from nose.tools import raises
 
 
@@ -56,6 +57,10 @@ def test_session():
         </xml>
     """
 
+    try:
+        os.remove(os.path.abspath("werobot_session"))
+    except OSError:
+        pass
     session_storages = [
         filestorage.FileStorage(),
         mongodbstorage.MongoDBStorage(pymongo.MongoClient().t.t),
@@ -65,9 +70,9 @@ def test_session():
     for session_storage in session_storages:
         remove_session(session_storage)
         robot.session_storage = session_storage
-        reply_1 = tester.send_xml(xml_1)
+        reply_1 = tester.send_xml(xml_1)._args['content']
         assert reply_1 == 'ss', (reply_1, session_storage)
-        reply_2 = tester.send_xml(xml_2)
+        reply_2 = tester.send_xml(xml_2)._args['content']
         assert reply_2 == 'ss', (reply_2, session_storage)
         remove_session(session_storage)