+++ travis/terryfy/__init__.py
@@ -1,0 +1,1 @@
+# Init for terryfy waf helpers