+++ werobot/exceptions.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, unicode_literals
+
+
+class ConfigError(Exception):
+    pass