+++ werobot/session/filestorage.py
@@ -2,12 +2,13 @@
 
 try:
     import anydbm as dbm
+
     assert dbm
 except ImportError:
     import dbm
 
 from werobot.session import SessionStorage
-from werobot.utils import json_loads, json_dumps
+from werobot.utils import json_loads, json_dumps, to_binary
 
 
 class FileStorage(SessionStorage):
@@ -16,8 +17,13 @@ class FileStorage(SessionStorage):
 
     :param filename: 文件名， 默认为 ``werobot_session``
     """
+
     def __init__(self, filename='werobot_session'):
-        self.db = dbm.open(filename, "c")
+        try:
+            self.db = dbm.open(filename, "c")
+        except TypeError:
+            # dbm in PyPy requires filename to be binary
+            self.db = dbm.open(to_binary(filename), "c")
 
     def get(self, id):
         try: